stafftext.ui update
Finally managed to sort this out! Aeolus stop list buttons are now of a
uniform colour and Tremulants have green text and couplers have red.
Had to resort to hacking the file in Notepad++ to get the width and
height back to the original values -Qt Designer inisits on increasing
the width to 1134 and then won't let you change it.


Former-commit-id: ede25022fb121908830ce9c52ea94713072a24c6