Add rendering of capo chords, and capo setting.
For now, the capo position is a ChordName style setting. In the
future, it can be enhanced to allow capo position to be changed
within a score, using something like TempoMap or TimeSigMap.
If capo position is non-zero, the chord to be played with capo
applied is shown in brackets after the standard chord.

Signed-off-by: lasconic <lasconic@gmail.com>

Former-commit-id: d316a4d9c84c789317541bc84e5954dcdf212862