TAB: letter shortcut for "fret-8" action changed from 'I' to 'J'
The letter shortcut for the "fret-8" action was 'I' and this created a conflict with the "instruments" action.

It has been changed to 'J' as:

*) J is not used by TAB and is them free
*) there is no conflict with "enharmonic-up", as enharmony makes no sense in TAB's and its actions are disabled in TAB note entry.

(Background: TAB note entry has double shortcuts for fret marks: a letter series and a digit series. TAB's can use both letters and numbers and having both shortcuts makes user life MUCH simpler)


Former-commit-id: 0d33762e82d3fb6451eecec48022c25df4cdfd1d