Updated SymId enum (sym.h) and Sym::synNames/symUserNames to SMuFL 0.7: - all SMuFL 0.7 entities (as names or symbol id) are listed - a few (15 in all) additional symbols are added to the end for    - SMuFL stylistic alternates MuseScore needs to access directly    - precomposed ornaments, SMuFL provides as sequences of glyphs    - glyphs SMuFL does not provides (yet?), but has been used as symbols
Updated mscore/glyphnames.json to match the above

Updated code points which were using SymId which have changed from SMuFL 0.6 to 0.7

Fixed note heads palette.

Note: The Dynamics palette is broken, as it accesses the mscore glyphs directly, but the palettes use Bravura. The usage of HTML texts in this palette is unclear: please advise.

Former-commit-id: d60f44cc0904e17567c81e4013995da71549c495