fix #23799: annotations deleted with chordrests
Segment annotations are currently deleted if the last chordrest in that
segment is deleted - either because you selected a region containing
this chordrest and hit Cut/Delete or because you lengthened a previous
chordrest to subsume this one.  But the 1.3, and expected, behavior is
that annotations remain even when chordrests are deleted.


Former-commit-id: 60599a3b4a004e3edea8dbbffff84e952aff2f5c