fix #23038: reduce height of dialogs to 560 pixel
so they easily fit a 1024x600 screen on e.g. a Netbook, without losing
anything, at least not on Windows.
The dialog editstyle.ui can only get reduced to 573px permanently (it
increases to that on the next change otherwise). It'd be nice if we could
reduce that dialog by the remain 13px, but isn't strictly needed,
with the window title added that dialog might be a few (single digit)
pixel higher than 600, but nothing relevant of that dialog should
become invisible, so that shouldn't be much of a problem.
The only remaining dialog, stafftype.ui, is taken care of in PR #763


Former-commit-id: 12497a46ec96e35150911ea8c1e124b0ee002c0b