Fix #25060 - Fix FiguredBass creating, editing and displaying.
After the recent changes to Text styling and editing, FiguredBass has been broken. This patch attempts to fix it.

As the current Text in-place editor insists in inserting spurious (and sometime wrong) HTML markup and styling which jeopardize the parsing of FiguredBassItem's, I have modified the FiguredBass endEdit() callback function to discard all HTML and keep only the plain text.

This may reduce the flexibility of entering extra indications in FiguredBass elements, but ensure that properly structured texts are parsed and recognized.

The patch also corrects two points in text.cpp which forced extra HTML formatting to be inserted.


Former-commit-id: 0a05faddb5416f3598cd6cd2d277e434117af233