Fix #22996 - Wrong clef when selecting standard staff type in instrument wizard.
In the instrument wizard, when a staff is changed to standard type (from a previous different type), the standard clef of the istrument is lost.

Changes:
- Added a member variable to `StaffListItem` storing the default standard clef for the instrument.
- Modified `instruments.xml' to remove &lt;clef&gt; tags from tablature instrument definitions: all these instruments are inited from a standard-staff main entry, recording the default standard clef; this clef is used to init the `StaffListItem::_defaultClef` member variable.

NOTE: Percussion instruments are not covered (an number of percussion instruments seem to have old-style definitions in `instruments.xml`); conversion from percussion to standard staff may yield unexpected results.

Former-commit-id: b8ecd4990556c341d73d33037e9d96f3de229912