Fix #25983 - Crash when copy-paste to TAB includes accidentals
The bug has been unveiled by this commit:
https://github.com/musescore/MuseScore/commit/1d6a4763780433b1f15c1c4b08845081be68e04e [formerly cbb388264af3ed1507b872cf7939a0611d36b5d0]
but it probably preexisted.

It is caused in function `Score::pasteStaff()` where the election status after the paste is updated before the re-layout and re-format process is completed (with a call to `updateNotes()`).

Among other things, laying out a TAB involves removing accidentals (which have no room in a TAB) and this leaves the selection with inconsistent contents.

Fixed by moving the selection update step at the end of the `Score::pasteStaff()` function, after the call to `updateNotes()`.


Former-commit-id: 0b7a7ac3671b8e0dee8db350fd67adcd143595a4