Add string data to 1.x scores.
Based on the MIDI program of the instrument, the code tries to add string data to instruments read from 1.x scores.

It is not fool-proof and covers only the most common (and less likely to have been customized) cases). For other cases, the user will have to change the part instrument or to copy the part to a suitably created instrument.

It covers:
- guitar(s)
- bass(es), either bass guitars and double bass
- violin, viola and cello, as their MIDI programs are used for a number of other string instruments; in these cases, string number and7or tuning may need to be customized, but this at least provides a departure point.

Comments are welcome.


Former-commit-id: 1ae0a84331e9a0869d15570c5fc11f164f091a16