Fix #29451 - No dots in TAB styles which use them
In a TAB with the "Stem through staff" style, augmentation dots are not shown.

Fixed, by invoking the function which creates dots for standard staff (`setDotY()`); this function is in a work-flow which is not executed for TAB's (as a number of layout steps, like those for accidentals, are not needed for TAB's).


Former-commit-id: 9594716f69bc39a9c9bdd398b4ce225b5c49c894