Add 4 lute fingering symbols to mscore font and to palette
This is in preparation of a more complete support for historic tablature to be added later.

The 4 symbols are added to the mscore font in the SMuFL code points. Currently, they are not correctly positioned, much like the other symbol-based fingering, i.e. the cello thumb position.

This is also to remind that further development of fingering layout will have to take into account that a number of fingerings are not really text but actually symbols.


Former-commit-id: b66bb2d751caf2bc944867b7dc0739de5b61544a