Fix #32261 : Local edit of bar line affects all bar lines
The use case detailed in the original issue (http://musescore.org/en/node/32261) was the following:

If edit mode is entered for a bar line with user-edit span (e.g. a tick) and then exited without any edit, as Ctrl has never been pressed, the edit is assumed to be global and the span of all the bar lines is changed to that specific bar line span.

Adds the following fixes:

- Entering and exiting bar line edit mode and doing nothing actually does nothing whatsoever.
- Any further span edit of a bar line whose span is already user-customized is assumed to be local to that bar line only. This makes using [Ctrl] for any further editing of a single bar line redundant.

Additionally:

- Any change in the span of system-initial bar line is always assumed to be local and no longer affects all the bar lines, as before.


Former-commit-id: aaabf64ae2be154d40fc419b86807b14775c355e