Adds to the Articulation class 4 symbols for lute tablature fingerings.
The Articulation class has been chosen because these symbols behave much like other articulations. The class already contains a number of symbols which are not really neither articulations or ornaments.

The new symbols have been grouped together with the somehow related cello thumb position symbol, at the end of the ArticulationType enum and set to `ArticulationShowIn::TABLATURE` only.

A specific case for them has been added in the `Chord::layoutArticulation()` method, as they do not follow the rules of other articulations.


Former-commit-id: 05b52e268170f4b6d792ac389a5ee567ae3da512