Entering lyrics underscore at score end
If a lyrics underscore is entered at score end (or where there is no place for a suitable new/next lyrics to move to):

- the last underscore gives no visual feedback, until the score is re-laid out
- the score view is left in a vague state (lyrics edit mode with no current lyrics)

Fixed by updating lyrics melisma display and leaving edit mode.

The lyrics whose melisma is being edited is also selected to give the user some 'cursor'.


Former-commit-id: 166756ee75c3dffaa50e9f3b456cb17403bf27e1