TAB: Make tabs aware of staff transpositions.
Feature has been requested and discussed at length in the forum thread: http://musescore.org/en/node/31016

To implement this, the fret-to-pitch relation has been made dependent of the staff, as each staff may have an individual transposition value in addition to the string data. And -- in addition -- on tick, in preparation of dealing with ottavas and capos.


Former-commit-id: 23f3348405f43cc63f9d8b7916d99e6234ab99b4