Fix #32966 - TAB: hide fingerings
Simple and cheap fix which simply hides `Fingering` elements in TAB's. No change to the linking / copying stuff: elements are still there, but generate no visible output in TAB. See issue http://musescore.org/en/node/32966 for details and discussion.

Should an option for hiding / showing these elements be added in the future, as discussed in the forum issue, the change would be trivial.

This does not affect lute-specific RH fingerings recently added, as luckily they are not real `Fingering`s, but `Articulation`s.


Former-commit-id: 507aa6cd01b81e177f748264e08c16ec9bcb23b8