Fix #46586 - Misplaced barline with large spatium
In `Barine::layout()`, `magS()` is passed to `layoutWidth()` to compute bar line width.

If bar lines are to be scaled, `layoutWidth()` uses it to multiply `spatiun()`, but `magS()` already takes `spatium()` into account and this results in `spatium()` being accounted for twice.

Fixed by using `mag()` instead in the call to `layoutWidth()`.


Former-commit-id: 6afc0573822dd3d4c69ad2c81a93fd79b91ecae8