Fix #48751 : Ottava not played back after undoing a change in span
As, when the span of an Ottava is changed, the staff it belongs to calls `Staff::updateOttava()` to rebuild its pitch offset map, expecting to find the Ottava spanners in the score, removing it before updating the property and re-adding it after leaves the staff with a wrong pitch offset map.

Those lines to remove and re-add the spanner have been added with this commit:
https://github.com/musescore/MuseScore/commit/2d235297c9c8f823e95636116c5b93d56d005460 [formerly 87ad7b4f25dc54e624f5e201b3400c59fdfdc1e0]
the reason for them is not clear.

Maps for Ottavas are rebuild from scratch at any change anyway, so perhaps those lines are not needed for Ottavas; advice from core developers is welcome!


Former-commit-id: 902565e25e34b245a17197545e3698f768000d47